removed nftlp_price context
