+++ src/pages/NFTs/launchpad/Launchpad.tsx
@@ -2,15 +2,14 @@ import React, { FC } from 'react'
 import { useRouteMatch, Route, Switch } from 'react-router-dom'
 import { LandingPage } from './pages/LandingPage'
 import { SingleCollection } from './pages/SingleCollection'
-import { NFTLaunchpadProvider, NFTLPSelectedProvider } from '../../../context/nft_launchpad'
-import { SolPriceProvider } from '../../../context/nftlp_price'
+import { NFTLaunchpadProvider, NFTLPSelectedProvider, USDCToggleProvider } from '../../../context/nft_launchpad'
 
 export const Launchpad: FC = () => {
   const { path } = useRouteMatch()
 
   return (
     <Switch>
-      <SolPriceProvider>
+      <USDCToggleProvider>
         <Route exact path={path}>
           <NFTLaunchpadProvider>
             <LandingPage />
@@ -21,7 +20,7 @@ export const Launchpad: FC = () => {
             <SingleCollection />
           </NFTLPSelectedProvider>
         </Route>
-      </SolPriceProvider>
+      </USDCToggleProvider>
     </Switch>
   )
 }
